// src/modules/information/interfaces/information.ts
export interface Information {
  integration_type_id: string;
  integration_type_name: string;
  hospital_id: string;
  hospital_name: string;
  organization_id: string;
  client_key: string;
  secret_key: string;
  created_at: string;
  created_by: string;
  updated_at: string;
  updated_by: string;
}